Module minerva_runonce::rabbitmq

This module wraps all functions related to operations that should be run once to prepare the message broker, when the entire system starts.

Functions

broker_spinlock

Awaits for message broker availability on a spinlock.

create_default_queues

Create default queues for a given virtual host on the message broker.

create_virtual_host

Creates a virtual host for the current tenant on the message broker.